AI/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
asttokens==2.4.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.9.7
beautifulsoup4==4.12.3
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==4.9.2
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.0.3
parso==0.8.4
pdf2docx==0.5.8
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
premailer==3.10.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.8.1
rarfile==4.2
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.7.10
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.6
urllib3==2.2.1
uvicorn==0.29.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.18.1
